Fix for Angular 7/8
