+++ angular/projects/spark-angular/schematics/collection.json
@@ -4,12 +4,6 @@
     "ng-add": {
       "description": "Add Spark Design System to an Angular project.",
       "factory": "./ng-add/index#ngAdd"
-    },
-    "ng-add-setup-project": {
-      "description": "Sets up the specified project after the ng-add dependencies have been installed.",
-      "private": true,
-      "factory": "./ng-add/setup-project",
-      "schema": "./ng-add/schema.json"
     }
   }
 }
